Move std::path unit tests to integration tests
