+++ library/std/benches/lib.rs
@@ -5,3 +5,4 @@
 extern crate test;
 
 mod hash;
+mod path;
